% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_nominatim}
\alias{geocode_nominatim}
\title{Lookup coordinates using the Nominatim OSM API}
\usage{
geocode_nominatim(address)
}
\arguments{
\item{address}{character string for the address}
}
\value{
tibble with results
}
\description{
This looks up location data (© OpenStreetMap contributors) including coordinates from the Nomination OSM API.
}
\details{
Data from this service is provided under the ODbL license which requires to share alike

See attribution and license here: \url{https://www.openstreetmap.org/copyright}

See usage policy here: \url{https://operations.osmfoundation.org/policies/nominatim/}
}
