% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_openalex}
\alias{geocode_openalex}
\title{Forward geocoding using OpenAlex API}
\usage{
geocode_openalex(orgname)
}
\arguments{
\item{orgname}{an organization or institute name}
}
\value{
tibble with orgname, ROR id etc with best scored data in first row
}
\description{
Uses OpenAlex API affiliation endpoint to lookup a orgname and return ROR ids
and coordinates
}
\details{
\url{https://ror.readme.io/docs/rest-api#affiliation-parameter}
}
\examples{
\dontrun{
if(interactive()){
 library(dplyr)
 geocode_openalex("KTH") \%>\%
   select(id, name, ends_with(c("lng", "lat")))
 }
}
}
