% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{institutions_cfg}
\alias{institutions_cfg}
\title{Config used internally}
\usage{
institutions_cfg()
}
\value{
list with two slots, one with the base_url of the dataset
one with the local download locations
}
\description{
This config holds the remote GRID dataset download url
and the corresponding local destination on disk
}
