% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{institutions_query}
\alias{institutions_query}
\title{Run a custom SQL query}
\usage{
institutions_query(sql_query, ...)
}
\arguments{
\item{sql_query}{the query}

\item{...}{other arguments to be passed to the sql() fcn}
}
\description{
Run a custom SQL query
}
