% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{institutions_search}
\alias{institutions_search}
\title{Full text search for names of institutions}
\usage{
institutions_search(search_term)
}
\arguments{
\item{search_term}{token query, phrase query or NEAR query (see http://www.sqlite.org/fts5.html)}
}
\value{
tibble with matching results
}
\description{
Full text search for names of institutions
}
\examples{
 institutions_search("^Stockholm AND University")
 institutions_search(tolower("Royal Institute of Technology"))
}
