% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{institutions_table}
\alias{institutions_table}
\title{Institution table}
\usage{
institutions_table(tablename)
}
\arguments{
\item{tablename}{the name of the table, which can be enumerated with institution_tables()}
}
\description{
Institution table
}
