% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{institutions_download}
\alias{institutions_download}
\title{Download the GRID dataset and install the local SQLite3 database}
\usage{
institutions_download(overwrite = FALSE, cfg = institutions_cfg())
}
\arguments{
\item{overwrite}{logical indicating if local db should be overwritten}

\item{cfg}{the config to use when downloading, by default institutions_cfg()}
}
\value{
logical indicating if the db exists locally, invisibly
}
\description{
Download the GRID dataset and install the local SQLite3 database
}
